+++ python/boomer/boosting/cpp/label_wise_statistics.cpp
@@ -41,11 +41,11 @@ DenseLabelWiseRefinementSearchImpl::~DenseLabelWiseRefinementSearchImpl() {
 void DenseLabelWiseRefinementSearchImpl::updateSearch(uint32 statisticIndex, uint32 weight) {
     // For each label, add the gradient and Hessian of the example at the given index (weighted by the given weight) to
     // the current sum of gradients and Hessians...
-    intp offset = statisticIndex * numLabels_;
+    uint32 offset = statisticIndex * numLabels_;
 
-    for (intp c = 0; c < numPredictions_; c++) {
-        intp l = labelIndices_ != NULL ? labelIndices_[c] : c;
-        intp i = offset + l;
+    for (uint32 c = 0; c < numPredictions_; c++) {
+        uint32 l = labelIndices_ != NULL ? labelIndices_[c] : c;
+        uint32 i = offset + l;
         sumsOfGradients_[c] += (weight * gradients_[i]);
         sumsOfHessians_[c] += (weight * hessians_[i]) ;
     }
@@ -62,7 +62,7 @@ void DenseLabelWiseRefinementSearchImpl::resetSearch() {
 
     // Reset the sum of gradients and Hessians for each label to zero and add it to the accumulated sums of gradients
     // and hessians...
-    for (intp c = 0; c < numPredictions_; c++) {
+    for (uint32 c = 0; c < numPredictions_; c++) {
         accumulatedSumsOfGradients_[c] += sumsOfGradients_[c];
         sumsOfGradients_[c] = 0;
         accumulatedSumsOfHessians_[c] += sumsOfHessians_[c];
@@ -123,13 +123,13 @@ void DenseLabelWiseStatisticsImpl::resetCoveredStatistics() {
 }
 
 void DenseLabelWiseStatisticsImpl::updateCoveredStatistic(uint32 statisticIndex, uint32 weight, bool remove) {
-    intp offset = statisticIndex * numLabels_;
+    uint32 offset = statisticIndex * numLabels_;
     float64 signedWeight = remove ? -((float64) weight) : weight;
 
     // For each label, add the gradient and Hessian of the example at the given index (weighted by the given weight) to
     // the total sums of gradients and Hessians...
-    for (intp c = 0; c < numLabels_; c++) {
-        intp i = offset + c;
+    for (uint32 c = 0; c < numLabels_; c++) {
+        uint32 i = offset + c;
         totalSumsOfGradients_[c] += (signedWeight * gradients_[i]);
         totalSumsOfHessians_[c] += (signedWeight * hessians_[i]);
     }
@@ -147,13 +147,13 @@ void DenseLabelWiseStatisticsImpl::applyPrediction(uint32 statisticIndex, Predic
     uint32 numPredictions = prediction->numPredictions_;
     const uint32* labelIndices = prediction->labelIndices_;
     const float64* predictedScores = prediction->predictedScores_;
-    intp offset = statisticIndex * numLabels_;
+    uint32 offset = statisticIndex * numLabels_;
 
     // Only the labels that are predicted by the new rule must be considered...
-    for (intp c = 0; c < numPredictions; c++) {
-        intp l = labelIndices != NULL ? labelIndices[c] : c;
+    for (uint32 c = 0; c < numPredictions; c++) {
+        uint32 l = labelIndices != NULL ? labelIndices[c] : c;
         float64 predictedScore = predictedScores[c];
-        intp i = offset + l;
+        uint32 i = offset + l;
 
         // Update the score that is currently predicted for the current example and label...
         float64 updatedScore = currentScores_[i] + predictedScore;
@@ -193,9 +193,9 @@ AbstractLabelWiseStatistics* DenseLabelWiseStatisticsFactoryImpl::create() {
     AbstractLabelWiseLoss* lossFunction = lossFunctionPtr_.get();
     AbstractRandomAccessLabelMatrix* labelMatrix = labelMatrixPtr_.get();
     // The number of examples
-    intp numExamples = labelMatrix->numExamples_;
+    uint32 numExamples = labelMatrix->numExamples_;
     // The number of labels
-    intp numLabels = labelMatrix->numLabels_;
+    uint32 numLabels = labelMatrix->numLabels_;
     // A matrix that stores the gradients for each example and label
     float64* gradients = (float64*) malloc(numExamples * numLabels * sizeof(float64));
     // A matrix that stores the Hessians for each example and label
@@ -203,11 +203,11 @@ AbstractLabelWiseStatistics* DenseLabelWiseStatisticsFactoryImpl::create() {
     // A matrix that stores the currently predicted scores for each example and label
     float64* currentScores = (float64*) malloc(numExamples * numLabels * sizeof(float64));
 
-    for (intp r = 0; r < numExamples; r++) {
-        intp offset = r * numLabels;
+    for (uint32 r = 0; r < numExamples; r++) {
+        uint32 offset = r * numLabels;
 
-        for (intp c = 0; c < numLabels; c++) {
-            intp i = offset + c;
+        for (uint32 c = 0; c < numLabels; c++) {
+            uint32 i = offset + c;
 
             // Calculate the initial gradient and Hessian for the current example and label...
             std::pair<float64, float64> pair = lossFunction->calculateGradientAndHessian(labelMatrix, r, c, 0);
